+++ DEVELOPMENT.md
@@ -1,5 +1,11 @@
 ## Development
 
+### Async and non-async
+
+[Unasync](https://unasync.readthedocs.io/en/latest/index.html) is used to generate non-async code based on the async implementation. Only make changes to files in `cerbos/sdk/_async`. Then run `./pw generate` to re-generate the code in `cerbos/sdk/_sync`.
+
+
+### Package management
 This project uses [PDM](https://pdm.fming.dev) with [Pyprojectx](https://pyprojectx.github.io) for package management. The `pw` script should be used when working with this project.
 
 Adding a new dependency
@@ -44,3 +50,4 @@ Publishing to PyPI
 # PyPI
 ./pw twine upload dist/*
 ```
+
